+++ tempest_fastapi_sdk/__init__.py
@@ -192,7 +192,7 @@ from tempest_fastapi_sdk.webpush import (
     WebPushSubscriptionSchema,
 )
 
-__version__: str = "0.29.1"
+__version__: str = "0.30.0"
 
 __all__: list[str] = [
     "CEP",